# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Comerc API Documentation",
        "description": "Documentação da API Comerc",
        "contact": {
            "email": "contato@leonardo-marques.tech"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/orders": {
            "get": {
                "tags": [
                    "Orders"
                ],
                "summary": "Lista todos os pedidos",
                "description": "Display a listing of the resource.",
                "operationId": "955bb77a45ae173a485fadb8befed93f",
                "responses": {
                    "200": {
                        "description": "Lista paginada de pedidos",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Order"
                                            }
                                        },
                                        "links": {
                                            "type": "object"
                                        },
                                        "meta": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Orders"
                ],
                "summary": "Cria um novo pedido",
                "description": "Store a newly created resource in storage.",
                "operationId": "2123ba58dadb22bd6f2c8081af562230",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/OrderCreateRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Pedido criado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Order"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Erro de validação"
                    }
                }
            }
        },
        "/api/orders/{id}": {
            "get": {
                "tags": [
                    "Orders"
                ],
                "summary": "Mostra detalhes de um pedido",
                "description": "Display the specified resource.",
                "operationId": "88d27387b1e409e075c18609d4a0c300",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do pedido",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detalhes do pedido",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Order"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Pedido não encontrado"
                    }
                }
            },
            "put": {
                "tags": [
                    "Orders"
                ],
                "summary": "Atualiza um pedido existente",
                "description": "Update the specified resource in storage.",
                "operationId": "38a1b85e8294795a1b0dc438a36e319f",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do pedido a ser atualizado",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/OrderCreateRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Pedido atualizado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Order"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Erro de validação",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "The given data was invalid."
                                        },
                                        "errors": {
                                            "type": "object",
                                            "example": {
                                                "client_id": [
                                                    "The client id field is required."
                                                ]
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Pedido não encontrado"
                    },
                    "500": {
                        "description": "Erro interno do servidor"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Orders"
                ],
                "summary": "Deleta um pedido",
                "description": "Remove the specified resource from storage.",
                "operationId": "bab875a1ed962fb33565ab9b48c0d494",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do pedido",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Pedido deletado com sucesso"
                    },
                    "404": {
                        "description": "Pedido não encontrado"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "Client": {
                "required": [
                    "id",
                    "name",
                    "email",
                    "phone",
                    "birthDate",
                    "address",
                    "neighborhood",
                    "zipCode"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "maxLength": 255,
                        "example": "João Silva"
                    },
                    "email": {
                        "type": "string",
                        "format": "email",
                        "maxLength": 255,
                        "example": "joao.silva@email.com"
                    },
                    "phone": {
                        "type": "string",
                        "maxLength": 20,
                        "example": "11999999999"
                    },
                    "birthDate": {
                        "type": "string",
                        "format": "date",
                        "example": "1985-12-25"
                    },
                    "address": {
                        "type": "string",
                        "maxLength": 255,
                        "example": "Rua Exemplo, 123"
                    },
                    "complement": {
                        "type": "string",
                        "maxLength": 255,
                        "example": "Apartamento 101",
                        "nullable": true
                    },
                    "neighborhood": {
                        "type": "string",
                        "maxLength": 255,
                        "example": "Centro"
                    },
                    "zipCode": {
                        "type": "string",
                        "maxLength": 9,
                        "example": "12345-678"
                    },
                    "createdAt": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-05-28T20:00:00Z",
                        "nullable": true
                    },
                    "updatedAt": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-05-29T15:30:00Z",
                        "nullable": true
                    }
                },
                "type": "object"
            },
...
            "Order": {
                "required": [
                    "id",
                    "client",
                    "products"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "client": {
                        "$ref": "#/components/schemas/Client"
                    },
                    "products": {
                        "type": "array",
                        "items": {
                            "required": [
                                "id",
                                "name",
                                "quantity",
                                "priceAtTimeOfOrder"
                            ],
                            "properties": {
                                "id": {
                                    "type": "integer",
                                    "example": 1
                                },
                                "name": {
                                    "type": "string",
                                    "example": "Produto A"
                                },
                                "type": {
                                    "type": "string",
                                    "example": "eletrônico",
                                    "nullable": true
                                },
                                "photoUrl": {
                                    "type": "string",
                                    "example": "products_photos/abcd1234.jpg",
                                    "nullable": true
                                },
                                "quantity": {
                                    "type": "integer",
                                    "example": 2
                                },
                                "priceAtTimeOfOrder": {
                                    "type": "number",
                                    "format": "float",
                                    "example": 39.98
                                }
                            },
                            "type": "object"
                        }
                    },
                    "createdAt": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-05-28T20:00:00Z",
                        "nullable": true
                    },
                    "updatedAt": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-05-29T15:30:00Z",
                        "nullable": true
                    }
                },
                "type": "object"
            },
            "OrderCreateRequest": {
                "required": [
                    "client_id",
                    "products"
                ],
                "properties": {
                    "client_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "products": {
                        "type": "array",
                        "items": {
                            "required": [
                                "product_id",
                                "quantity"
                            ],
                            "properties": {
                                "product_id": {
                                    "type": "integer",
                                    "example": 1
                                },
                                "quantity": {
                                    "type": "integer",
                                    "minimum": 1,
                                    "example": 2
                                }
                            },
                            "type": "object"
                        },
                        "minItems": 1
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}